requirements.txt
playwright==1.43.0
python-dotenv==1.0.1
pymongo==4.7.2
pandas==2.2.2
openpyxl==3.1.2
tenacity==8.3.0
rich==13.7.1
aiofiles==23.2.1
pypdfium2==4.30.1
Pillow==10.4.0
